requirements.txt
pika==1.2.0
psycopg2-binary==2.9.3
